+++ src/main/java/com/veribadge/veribadge/config/SecurityConfig.java
@@ -1,11 +1,10 @@
 package com.veribadge.veribadge.config;
 
 import com.veribadge.veribadge.handler.CustomOAuth2SuccessHandler;
-import com.veribadge.veribadge.jwt.JwtAuthenticationFilter;
+import com.veribadge.veribadge.security.jwt.JwtAuthenticationFilter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.http.SessionCreationPolicy;
@@ -14,7 +13,6 @@ import org.springframework.security.web.authentication.UsernamePasswordAuthentic
 import org.springframework.web.cors.CorsConfiguration;
 
 
-import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
