+++ src/main/java/com/veribadge/veribadge/config/mock/VerificationInitializer.java
@@ -1,9 +1,9 @@
-package com.veribadge.veribadge.mock;
+package com.veribadge.veribadge.config.mock;
 
 import com.veribadge.veribadge.domain.Member;
 import com.veribadge.veribadge.domain.Verification;
 import com.veribadge.veribadge.exception.CustomException;
-import com.veribadge.veribadge.global.status.ErrorStatus;
+import com.veribadge.veribadge.common.status.ErrorStatus;
 import com.veribadge.veribadge.repository.MemberRepository;
 import com.veribadge.veribadge.repository.VerificationRepository;
 import lombok.RequiredArgsConstructor;
