+++ src/main/java/com/veribadge/veribadge/controller/BadgeController.java
@@ -3,7 +3,7 @@ package com.veribadge.veribadge.controller;
 import com.veribadge.veribadge.dto.BadgeVerifyResponseDto;
 import com.veribadge.veribadge.dto.CommentTagVerifyRequestDto;
 import com.veribadge.veribadge.exception.Response;
-import com.veribadge.veribadge.global.status.SuccessStatus;
+import com.veribadge.veribadge.common.status.SuccessStatus;
 import com.veribadge.veribadge.service.BadgeMatcherService;
 import io.swagger.v3.oas.annotations.Operation;
 import jakarta.validation.Valid;
