+++ src/main/java/com/veribadge/veribadge/handler/CustomOAuth2SuccessHandler.java
@@ -3,13 +3,10 @@ package com.veribadge.veribadge.handler;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.veribadge.veribadge.domain.Member;
 import com.veribadge.veribadge.domain.enums.Role;
-import com.veribadge.veribadge.dto.LoginResponseDto;
 import com.veribadge.veribadge.exception.CustomException;
-import com.veribadge.veribadge.exception.Response;
-import com.veribadge.veribadge.global.status.ErrorStatus;
-import com.veribadge.veribadge.global.status.SuccessStatus;
-import com.veribadge.veribadge.jwt.JwtGoogleProvider;
-import com.veribadge.veribadge.jwt.JwtKakaoProvider;
+import com.veribadge.veribadge.common.status.ErrorStatus;
+import com.veribadge.veribadge.security.jwt.JwtGoogleProvider;
+import com.veribadge.veribadge.security.jwt.JwtKakaoProvider;
 import com.veribadge.veribadge.repository.MemberRepository;
 import com.veribadge.veribadge.service.social.GoogleService;
 import com.veribadge.veribadge.service.MyBadgeService;
