+++ src/main/java/com/veribadge/veribadge/veribadge_extension/images/gold.png
