+++ src/main/java/com/veribadge/veribadge/veribadge_extension/images/platinum.png
