+++ src/main/java/com/veribadge/veribadge/veribadge_extension/injectBadge.js
